+++ ui/src/components/under-construction/index.tsx
@@ -2,26 +2,14 @@
 // Copyright (C) 2018-2019 Prusa Research s.r.o. - www.prusa3d.com
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-import { h, Fragment } from "preact";
-import Title from "../../components/title"
-import under_construction from "../../assets/under_construction.gif"
-
-interface P {
+import { h } from "preact";
+const UnderConstruction = props => {
+    if (process.env.PRINTER != "Original Prusa Mini") {
+        return require('./component').default(props);
+    } else {
+        return (<div />)
+    }
 }
-
-const UnderConstruction: preact.FunctionalComponent<P> = props => {
-    return (
-        <Fragment>
-            <Title id="under-construction" default_text="Under construction" />
-            <div class="columns">
-                <div class="column is-4 is-offset-4">
-                    <img src={under_construction} />
-                </div>
-            </div>
-        </Fragment>
-    );
-}
-
 
 export default UnderConstruction;
