+++ ui/src/routes/home/index.tsx
@@ -31,10 +31,7 @@ export const Home: preact.FunctionalComponent<homeProps> = props => {
         <div class="column">
           <StatusProgress {...props.progress_bar} />
           <br />
-          {
-            process.env.PRINTER != "Original Prusa Mini" &&
-            <Upload url={null} path={null} />
-          }
+          <Upload url={null} path={null} />
           <br />
           <Temperature temperatures={props.temperatures} bigSize={false} />
         </div>
